REF: Move linked_attribute and link_signal_to_widget from positioner to utils as it can be useful to more screens and templates.
