+++ typhos/utils.py
@@ -3,10 +3,12 @@ Utility functions for typhos
 """
 import collections
 import contextlib
+import functools
 import importlib.util
 import inspect
 import io
 import logging
+import operator
 import os
 import pathlib
 import random
@@ -23,6 +25,8 @@ import ophyd.sim
 from ophyd import Device
 from ophyd.signal import EpicsSignalBase, EpicsSignalRO
 from pydm.exception import raise_to_operator  # noqa
+
+from typhos import plugins
 
 try:
     import happi
@@ -1114,3 +1118,75 @@ def widget_to_image(widget, fill_color=QtCore.Qt.transparent):
     widget.render(image)
     painter.end()
     return image
+
+
+def link_signal_to_widget(signal, widget):
+    """
+    Registers the signal with PyDM, and sets the widget channel.
+
+    Parameters
+    ----------
+    signal : ophyd.OphydObj
+        The signal to use.
+
+    widget : QtWidgets.QWidget
+        The widget with which to connect the signal.
+    """
+    if signal is not None:
+        plugins.register_signal(signal)
+        if widget is not None:
+            widget.channel = channel_from_signal(signal)
+
+
+def linked_attribute(property_attr, widget_attr):
+    """
+    Decorator which connects a device signal with a widget.
+
+    Retrieves the signal from the device, registers it with PyDM, and sets the
+    widget channel.
+
+    Parameters
+    ----------
+    property_attr : str
+        This is one level of indirection, allowing for the component attribute
+        to be configurable by way of designable properties.
+        In short, this looks like:
+            ``getattr(self.device, getattr(self, property_attr))``
+        The component attribute name may include multiple levels (e.g.,
+        ``'cpt1.cpt2.low_limit'``).
+
+    widget_attr : str
+        The attribute name of the widget, referenced from ``self``.
+        The component attribute name may include multiple levels (e.g.,
+        ``'ui.low_limit'``).
+    """
+    get_widget_attr = operator.attrgetter(widget_attr)
+
+    def wrapper(func):
+        @functools.wraps(func)
+        def wrapped(self):
+            widget = get_widget_attr(self)
+            device_attr = getattr(self, property_attr)
+            get_device_attr = operator.attrgetter(device_attr)
+
+            try:
+                signal = get_device_attr(self.device)
+            except AttributeError:
+                signal = None
+            else:
+                # Fall short of an `isinstance(signal, OphydObj) check here:
+                try:
+                    link_signal_to_widget(signal, widget)
+                except Exception:
+                    logger.exception(
+                        'device.%s => self.%s (signal: %s widget: %s)',
+                        device_attr, widget_attr, signal, widget)
+                    signal = None
+                else:
+                    logger.debug('device.%s => self.%s (signal=%s widget=%s)',
+                                 device_attr, widget_attr, signal, widget)
+
+            return func(self, signal, widget)
+
+        return wrapped
+    return wrapper
